tests have been re structured
